Feat(UI): Add dark mode toggle and make responsive
Added a dark mode toggle and refactored the stylesheet to use variables instead of hardcoded values for select properties. Two lines of instructions on how to rotate and move the ships were also added. The boards will also stack on top of each other on smaller screens, and won't lose their aspect ratio.

- Updated `setupInterface()` to display instructions before starting the game
- Added `removeInstructions()` to remove instructions at game start
- Added `addDarkModeBtn()` and `toggleDarkMode()` to switch themes
- Refactored `styles.css` to define multiple values as variables to facilitate creating a dark theme